Fix "can't use env.foo lookups in django templates" error:
This is caused by django's template system looking up 'env' seeing
that it is callable, calling it, and then trying to look up '.foo' of
the result. Due to the 'context runner' feature in tav's fork, env is
actually callable, which will break this.

Thus, if env is called with no arguments (the way it will be during
django template resolution), return self, so that 'env.foo' works as
expected in django templates